Move CheckChannelShutdown command from channel to network
