+++ crates/fiber-lib/src/fiber/channel.rs
@@ -4,15 +4,12 @@ use super::{
     gossip::SOFT_BROADCAST_MESSAGES_CONSIDERED_STALE_DURATION, graph::ChannelUpdateInfo,
     types::ForwardTlcResult,
 };
-use crate::ckb::GetShutdownTxRequest;
-use crate::ckb::GetShutdownTxResponse;
 use crate::fiber::config::MILLI_SECONDS_PER_EPOCH;
 use crate::fiber::fee::{check_open_channel_parameters, check_tlc_delta_with_epochs};
 #[cfg(any(debug_assertions, feature = "bench"))]
 use crate::fiber::network::DebugEvent;
 use crate::fiber::network::PaymentCustomRecords;
 use crate::{debug_event, fiber::types::TxAbort, utils::tx::compute_tx_message};
-use ckb_types::core::tx_pool::TxStatus;
 #[cfg(test)]
 use musig2::BinaryEncoding;
 use musig2::SecNonceBuilder;
@@ -126,9 +123,6 @@ pub const PEER_CHANNEL_RESPONSE_TIMEOUT: u64 = 10 * 1000;
 // multiple 5 times for max tlc number of tlcs is a safe number
 const MAX_RETRYABLE_TLC_OPERATIONS: u16 = SYS_MAX_TLC_NUMBER_IN_FLIGHT as u16 * 5;
 
-// The interval time to check shutdown transaction
-const CHANNEL_CHECK_SHUTDOWN_INTERVAL: Duration = Duration::from_secs(300);
-
 #[derive(Debug)]
 pub enum ChannelActorMessage {
     /// Command are the messages that are sent to the channel actor to perform some action.
@@ -175,7 +169,6 @@ pub enum ChannelCommand {
     ForwardTlcResult(ForwardTlcResult),
     #[cfg(any(test, feature = "bench"))]
     ReloadState(ReloadParams),
-    CheckShutdown,
 }
 
 impl Display for ChannelCommand {
@@ -191,7 +184,6 @@ impl Display for ChannelCommand {
             ChannelCommand::ForwardTlcResult(res) => write!(f, "ForwardTlcResult [{:?}]", res),
             #[cfg(any(test, feature = "bench"))]
             ChannelCommand::ReloadState(_) => write!(f, "ReloadState"),
-            ChannelCommand::CheckShutdown => write!(f, "CheckShutdown"),
         }
     }
 }
@@ -2170,59 +2162,6 @@ where
                 }
                 Ok(())
             }
-            ChannelCommand::CheckShutdown => {
-                // check channel ready state
-                if state.state == ChannelState::ChannelReady {
-                    // check shutdown transactions
-                    let request = GetShutdownTxRequest {
-                        funding_lock_script: state.get_funding_lock_script(),
-                    };
-                    match call!(self.network, |tx| NetworkActorMessage::Command(
-                        NetworkActorCommand::GetShutdownTx(request, tx)
-                    ))
-                    .expect(ASSUME_NETWORK_ACTOR_ALIVE)
-                    {
-                        Ok(Some(GetShutdownTxResponse {
-                            transaction: Some(tx),
-                            tx_status: TxStatus::Committed(..),
-                        })) => {
-                            // we only check remote sent force close transaction here
-                            if tx.outputs().len() == 1 {
-                                if let Some(output) = tx.outputs().get(0) {
-                                    // Check if channel is force closed by counter party
-                                    let lock_args =
-                                        &blake2b_256(state.get_commitment_lock_script_xonly(true))
-                                            [0..20];
-                                    if &output.lock().args().raw_data()[0..20] == lock_args {
-                                        let channel_id = state.get_id();
-                                        let peer_id = state.get_remote_peer_id();
-                                        let tx_hash = tx.hash();
-                                        self.network
-                                            .send_message(NetworkActorMessage::Event(
-                                                NetworkActorEvent::ClosingTransactionConfirmed(
-                                                    peer_id, channel_id, tx_hash, true, false,
-                                                ),
-                                            ))
-                                            .expect(ASSUME_NETWORK_ACTOR_ALIVE);
-                                    }
-                                }
-                            }
-                        }
-                        Ok(_) => {
-                            // no shutdown tx
-                        }
-                        Err(err) => {
-                            warn!(
-                                "Error during get channel shutdown tx {} from peer {:?} {:?}",
-                                state.get_id(),
-                                state.get_remote_peer_id(),
-                                err
-                            );
-                        }
-                    }
-                }
-                Ok(())
-            }
         }
     }
 
@@ -2831,11 +2770,6 @@ where
                 }
             }
         }
-
-        // check channel shutdown
-        myself.send_interval(CHANNEL_CHECK_SHUTDOWN_INTERVAL, || {
-            ChannelActorMessage::Command(ChannelCommand::CheckShutdown)
-        });
 
         Ok(())
     }
@@ -7633,7 +7567,7 @@ impl ChannelActorState {
         }
     }
 
-    fn get_commitment_lock_script_xonly(&self, for_remote: bool) -> [u8; 32] {
+    pub(crate) fn get_commitment_lock_script_xonly(&self, for_remote: bool) -> [u8; 32] {
         let local_pubkey = self.get_local_channel_public_keys().funding_pubkey;
         let remote_pubkey = self.get_remote_channel_public_keys().funding_pubkey;
         let pubkeys = if for_remote {
